requirements.txt
fastapi==0.112.0
uvicorn==0.15.0
python-multipart==0.0.5
torch==1.9.0
torchvision==0.10.0
transformers==4.10.0
Pillow==8.3.1
pinecone-client==2.1.0
PyMuPDF==1.18.15
python-dotenv==0.19.0